requirements.txt
cffi==2.0.0
cryptography==46.0.5
pycparser==3.0
PyYAML==6.0.3
pytest==8.4.2
